2022/10/02 16:18

+++ x2.rs
@@ -8,7 +8,7 @@ fn main(){
     }
 }
 
-fn x2(num: i32) -> i32 {
+fn x2(num: u64) -> u64 {
     num * 2
 }
